+++ cmd/redact/runtime.go
@@ -7,8 +7,8 @@ import (
 	"strings"
 
 	"github.com/julian7/redact/files"
+	"github.com/julian7/redact/logger"
 	"github.com/julian7/redact/sdk"
-	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
@@ -17,7 +17,7 @@ import (
 type cmdFactory func() (*cobra.Command, error)
 
 type Runtime struct {
-	*logrus.Logger
+	*logger.Logger
 	*files.MasterKey
 	*viper.Viper
 	Config string
@@ -128,26 +128,12 @@ func (rt *Runtime) SetupLogging(cmd *cobra.Command, args []string) {
 }
 
 func (rt *Runtime) setLogLevel(level string) {
-	var logLevel logrus.Level
+	err := rt.Logger.SetLevelFromString(level)
+	if err != nil {
+		rt.Logger.Warnf("cannot set log level: %v", err)
 
-	switch level {
-	case "debug":
-		logLevel = logrus.DebugLevel
-	case "info":
-		logLevel = logrus.InfoLevel
-	case "warn":
-		logLevel = logrus.WarnLevel
-	case "error":
-		logLevel = logrus.ErrorLevel
-	case "fatal":
-		logLevel = logrus.FatalLevel
-	case "":
-		return
-	default:
-		rt.Logger.Warnf("unknown log level: %s", level)
 		return
 	}
 
-	rt.Logger.SetLevel(logLevel)
 	rt.Logger.Debugf("Setting log level to %s", level)
 }