+++ files/global_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/julian7/redact/files"
 	"github.com/julian7/redact/gitutil"
-	"github.com/sirupsen/logrus"
+	"github.com/julian7/redact/logger"
 	"github.com/spf13/afero"
 )
 
@@ -18,7 +18,7 @@ const (
 func genGitRepo() (*files.MasterKey, error) {
 	k := &files.MasterKey{
 		Fs:     afero.NewMemMapFs(),
-		Logger: logrus.New(),
+		Logger: logger.New(),
 		RepoInfo: gitutil.GitRepoInfo{
 			Common:   ".git",
 			Toplevel: "/git/repo",