+++ logger/doc.go
@@ -1,0 +1,2 @@
+// Logger is a minimalistic leveled log writer based on log package.
+package logger